Work with design. Add splash activity

+++ app/src/main/java/com/example/weatherapp/dataClasses/Coord.kt
@@ -1,6 +1,6 @@
 package com.example.weatherapp.dataClasses
 
 data class Coord(
-    val lat: Int,
+    val lat: Double,
     val lon: Double
 )
